pico5000 driver drafted

+++ src/honeychrome/trace_analyst.py
@@ -269,7 +269,7 @@ if __name__ == '__main__':
     mp.set_start_method("spawn")
 
     # Allocate shared memory block, plus head and tail indices
-    traces_cache_shm = shared_memory.SharedMemory(create=True, size=np.zeros(traces_cache_size, dtype=dtype).nbytes)
+    traces_cache_shm = shared_memory.SharedMemory(create=True, size=np.zeros(traces_cache_size, dtype=np.uint16).nbytes)
     traces_cache_lock = Lock()
     index_head_traces_cache = mp.Value('i', 0)
     index_tail_traces_cache = mp.Value('i', 0)
